requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.7.0
pymssql==2.2.7
PyMySQL==1.0.3
pyodbc==4.0.39
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
typing-extensions==4.6.2
urllib3==2.0.4
uuid==1.30
Werkzeug==2.2.3
wfastcgi==3.0.0
zipp==3.15.0